UI Setting

+++ Test_SkillContest/Assets/SkillContest/Scripts/EnemyData.cs
@@ -1,9 +1,9 @@
 [System.Serializable]
 public class EnemyData
 {
-    public float SpawnDelay;
+    public int SpawnDelay;
     public int SpawnPos;
     public string SpawnType;
     public int MoveDir;
-    public float WaitTime;
+    public int WaitTime;
 }
